erros ajustados

+++ src/services/products/deleteProduct.service.js
@@ -2,8 +2,20 @@ import database from "../../database";
 
 const deleteProductService = async (id) => {
     try {
+        const exist = await database.query(`SELECT * FROM products WHERE products.id = $1`, [id]);
+
+        if(exist.rows.length == 0) {
+            return {
+                message: 'Product not found',
+                statusCode: 404
+            }
+        }
+
         const res = await database.query(`DELETE FROM products WHERE products.id = $1`, [id]);
-        return `Product deleted.`
+        return {
+            message: '',
+            statusCode: 204
+        };
     } catch (error) {
         throw new Error(error);
     }
